requirements.txt
opencv_python
tensorflow
requests
pytz
beautifulsoup4
numpy
Pillow
pycryptodome
PyYAML
